Reorganise files, pending remake of frontend
