+++ client/src/models/Dropper.js
@@ -3,9 +3,136 @@
 // models/
 // Dropper.js
 
-import { Peer, FileStream } from './index.js';
+import * as uuid from 'uuid';
+
+import { Peer } from './index.js';
 
 const _messageSize = parseInt(process.env.REACT_APP_MESSAGE_SIZE, 10);
+
+/* FileStream - Send a file through a data channel.
+ *
+ * public methods:
+ *
+ * constructor(peer, file) - Start a drop for a given file.
+ *   peer: the peer connection to facilitate the drop
+ *   file: the file to drop
+ *
+ * public fields:
+ *
+ * label - The data channel label
+ *
+ * offset - Current position in file (bytes)
+ *
+ * dispatches events:
+ *
+ * 'done' -> Event
+ * - The file was sent successfully.
+ */
+class FileStream extends EventTarget {
+  // private fields
+
+  _dataChannel = null; // the data channel
+  _file = null; // the file to be sent
+  _state = 'fileinfo'; // the current state
+
+  // public fields
+
+  label; // the data channel label
+  offset = 0; // the current position in the file
+
+  // constructor
+
+  constructor(peer, file) {
+    super();
+
+    // generate a unique UUID for this file
+    this.label = uuid.v4();
+
+    this._dataChannel = peer.createDataChannel(this.label);
+    this._file = file;
+
+    // event listeners for data channel
+    this._dataChannel.addEventListener('open', this._sendMessage.bind(this));
+    this._dataChannel.addEventListener(
+      'bufferedamountlow',
+      this._sendMessage.bind(this)
+    );
+  }
+
+  // private methods
+
+  async _sendMessage() {
+    try {
+      switch (this._state) {
+        // send file information to the recipient
+        case 'fileinfo':
+          const packet = {
+            type: 'fileinfo',
+            fileinfo: {
+              name: this._file.name,
+              size: this._file.size,
+              type: this._file.type
+            }
+          };
+
+          this._dataChannel.send(JSON.stringify(packet));
+          this._state = 'send';
+
+          break;
+
+        // ready to send next chunk
+        case 'send':
+          // check if done sending file
+          if (this.offset >= this._file.size) {
+            this._dataChannel.send('{"type":"done"}'); // send done message
+            // let recipient close the connection
+
+            // set state to done
+            this._state = 'done';
+
+            return;
+          }
+
+          // incase an event follows the 'await' below, don't send
+          // (preserve the proper ordering of array buffers)
+          this._state = 'sending';
+
+          // slice the file given the current position
+          let end = Math.min(this.offset + _messageSize, this._file.size);
+          let blob = this._file.slice(this.offset, end);
+
+          // update the current position
+          this.offset = end;
+
+          // get buffer and send it through the data channel
+          // NOTE: this._dataChannel will dispatch bufferedamountlow when done
+          let buffer = await blob.arrayBuffer();
+          this._dataChannel.send(buffer);
+
+          // next event should send
+          this._state = 'send';
+
+          // take out the trash
+          buffer = null;
+          blob = null;
+
+          break;
+
+        // 'sending', 'done', ...
+        default:
+          console.log(
+            `FileStream: Passing state '${this._state}' in _sendMessage.`
+          );
+
+          // pass
+
+          break;
+      }
+    } catch (err) {
+      console.log(`FileStream: Error in _sendMessage: ${err.toString()}`);
+    }
+  }
+}
 
 /* Dropper - Send files.
  *
@@ -61,12 +188,6 @@ export class Dropper extends EventTarget {
     for (let i = 0; i < files.length; i++) {
       const fileStream = new FileStream(this._peer, files[i]);
 
-      fileStream.addEventListener('offsetchanged', (event) => {
-        this.dispatchEvent(new MessageEvent('offsetchanged', {
-          data: { label: fileStream.label, offset: event.data }
-        }));
-      });
-
       this.fileinfo[fileStream.label] = {
         name: files[i].name,
         size: files[i].size,
@@ -93,6 +214,12 @@ export class Dropper extends EventTarget {
 
     // watch file streams, and, when finished
     this._awaitFileStreams();
+
+    this._interval = setInterval(() => {
+      this.dispatchEvent(new MessageEvent('offsetchanged',
+        { data: this._fileStreams.map((fileStream) => fileStream.offset) }
+      ));
+    }, 1000); // every second
   }
 
   // private methods
