+++ client/src/models/index.js
@@ -4,8 +4,4 @@
 // index.js
 
 export * from './Dropper.js';
-export * from './FileStore.js';
-export * from './FileStream.js';
-export * from './Peer.js';
 export * from './Recipient.js';
-export * from './SignalChannel.js';
